+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 Written with guidance from [Keep a CHANGELOG](http://keepachangelog.com/).
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- parse will take a list of strings
+- JSONPointer.extract returns a list of paths that make up an object 
+
 ## v1.1.0
 ## Added
 - pointers can be passed as lists as well as strings